Выносит функции для рендеринга
