+++ src/components/event-offer.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render.js';
 
 const createEventOfferTemplate = (offer) => {
   const {title, price} = offer;
