+++ src/components/form-destination.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render.js';
 
 const createDestinationTemplate = (destination) => {
   return (
