+++ src/components/offer-photos.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render.js';
 
 const createPhotosTemplate = (photoUrl) => {
   return (
