+++ src/components/offer.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render.js';
 
 const createOfferTemplate = (offer) => {
   const {type, title, price} = offer;
