+++ src/components/trip-cost.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/render.js';
 
 const createTripCostTemplate = (cost) => {
   return (
